+++ scripts/generate-testkit-exports/testkit-definitions.js
@@ -77,9 +77,19 @@ module.exports = {
 
   DragDropContextProvider: { skipSanityTest, noTestkit },
 
-  EndorseContentLayout: { skipSanityTest },
+  EndorseContentLayout: {
+    skipSanityTest,
+    // TODO: testkit of this component is not following convention.
+    // It is used only internally and before automated export was not exported manually
+    noTestkit,
+  },
 
-  GoogleAddressInput: { skipSanityTest },
+  GoogleAddressInput: {
+    skipSanityTest,
+    // TODO: testkit of this component is not following convention.
+    // It is used only internally and before automated export was not exported manually
+    noTestkit,
+  },
 
   GoogleAddressInputWithLabel: { skipSanityTest },
 
@@ -90,22 +100,6 @@ module.exports = {
   Layout: { skipSanityTest, noTestkit },
 
   MessageBox: { skipSanityTest, noTestkit },
-
-  ButtonHeader: {
-    // it's actually Card.ButtonHeader, should be deprecated
-    testkitPath: '../src/Card/ButtonHeader/ButtonHeader.driver',
-    skipSanityTest,
-  },
-
-  LinkHeader: {
-    testkitPath: '../src/Card/LinkHeader/LinkHeader.driver',
-    skipSanityTest,
-  },
-
-  CollapsedHeader: {
-    skipSanityTest,
-    testkitPath: '../src/Card/CollapsedHeader/CollapsedHeader.driver',
-  },
 
   Header: {
     testkitPath: '../src/Card/Header/Header.driver',
@@ -207,11 +201,15 @@ module.exports = {
 
   BadgeSelect: {},
 
-  CalendarPanel: {},
-
   Breadcrumbs: {},
 
   Calendar: {},
+
+  CalendarPanel: {},
+
+  CalendarPanelFooter: {
+    unidriver,
+  },
 
   DataTable: {},
 
@@ -272,33 +270,8 @@ module.exports = {
 
   DropdownBase: { unidriver },
 
-  TpaLabel: {
-    testkitPath: '../src/TPA/Label/Label.driver',
-    skipSanityTest,
-  },
-
-  TpaTextLink: {
-    testkitPath: '../src/TPA/TextLink/TextLink.driver',
-    skipSanityTest,
-  },
-
-  TpaButton: {
-    skipSanityTest,
-    testkitPath: '../src/TPA/Button/Button.driver',
-  },
-
-  TpaFloatingTabs: {
-    testkitPath: '../src/TPA/FloatingTabs/FloatingTabs.driver',
-    skipSanityTest,
-  },
-
   RadioButton: {
     testkitPath: '../src/RadioGroup/RadioButton/RadioButton.driver',
-    skipSanityTest,
-  },
-
-  TpaInput: {
-    testkitPath: '../src/TPA/Input/Input.driver',
     skipSanityTest,
   },
 
@@ -316,4 +289,6 @@ module.exports = {
     testkitPath: '../src/BaseComponents/TextLinkLayout/TextLinkLayout.driver',
     skipSanityTest,
   },
+
+  Box: { unidriver },
 };